+++ test/test_formatters.rb
@@ -28,6 +28,14 @@ class FormatterTest < TwineTest
     end
   end
 
+  def assert_nested_translations_read_correctly
+    1.upto(4) do |i|
+      assert_equal "value#{i}-english", @empty_twine_file.definitions_by_key["key#{i}"].translations['en']
+    end
+    assert_equal "value5a-english", @empty_twine_file.definitions_by_key["key5.key5a"].translations['en']
+    assert_equal "value5b-english", @empty_twine_file.definitions_by_key["key5.key5b"].translations['en']
+  end
+
   def assert_file_contents_read_correctly
     assert_translations_read_correctly
 
@@ -47,26 +55,87 @@ class TestAndroidFormatter < FormatterTest
       'a "good" way'              => 'a \"good\" way',
 
       '<b>bold</b>'               => '<b>bold</b>',
+      '<em>bold</em>'             => '<em>bold</em>',
+
       '<i>italic</i>'             => '<i>italic</i>',
+      '<cite>italic</cite>'       => '<cite>italic</cite>',
+      '<dfn>italic</dfn>'         => '<dfn>italic</dfn>',
+
+      '<big>larger</big>'         => '<big>larger</big>',
+      '<small>smaller</small>'    => '<small>smaller</small>',
+
+      '<font color="#45C1D0">F</font>' => '<font color="#45C1D0">F</font>',
+
+      '<tt>monospaced</tt>'       => '<tt>monospaced</tt>',
+
+      '<s>strike</s>'             => '<s>strike</s>',
+      '<strike>strike</strike>'   => '<strike>strike</strike>',
+      '<del>strike</del>'         => '<del>strike</del>',
+
       '<u>underline</u>'          => '<u>underline</u>',
 
+      '<super>superscript</super>'=> '<super>superscript</super>',
+
+      '<sub>subscript</sub>'      => '<sub>subscript</sub>',
+
+      '<ul>bullet point</ul>'     => '<ul>bullet point</ul>',
+      '<li>bullet point</li>'     => '<li>bullet point</li>',
+
+      '<br>line break'            => '<br>line break',
+
+      '<div>division</div>'       => '<div>division</div>',
+
+      '<span style="color:#45C1D0">inline</span>' => '<span style="color:#45C1D0">inline</span>',
+
+      '<p>para</p>'               => '<p>para</p>',
+      '<p dir="ltr">para</p>'     => '<p dir="ltr">para</p>',
+
       '<b>%@</b>'                 => '&lt;b>%s&lt;/b>',
+      '<em>%@</em>'               => '&lt;em>%s&lt;/em>',
+
       '<i>%@</i>'                 => '&lt;i>%s&lt;/i>',
+      '<cite>%@</cite>'           => '&lt;cite>%s&lt;/cite>',
+      '<dfn>%@</dfn>'             => '&lt;dfn>%s&lt;/dfn>',
+
+      '<big>%@</big>'             => '&lt;big>%s&lt;/big>',
+      '<small>%@</small>'         => '&lt;small>%s&lt;/small>',
+
+      '<font color="#45C1D0>%@</font>' => '&lt;font color="#45C1D0>%s&lt;/font>',
+
+      '<tt>%@</tt>'               => '&lt;tt>%s&lt;/tt>',
+
+      '<s>%@</s>'                 => '&lt;s>%s&lt;/s>',
+      '<strike>%@</strike>'       => '&lt;strike>%s&lt;/strike>',
+      '<del>%@</del>'             => '&lt;del>%s&lt;/del>',
+
       '<u>%@</u>'                 => '&lt;u>%s&lt;/u>',
 
-      '<span>inline</span>'       => '&lt;span>inline&lt;/span>',
-      '<p>paragraph</p>'          => '&lt;p>paragraph&lt;/p>',
+      '<super>%@</super>'         => '&lt;super>%s&lt;/super>',
+
+      '<sub>%@</sub>'             => '&lt;sub>%s&lt;/sub>',
+
+      '<ul>%@</ul>'               => '&lt;ul>%s&lt;/ul>',
+      '<li>%@</li>'               => '&lt;li>%s&lt;/li>',
+
+      '<br>%@'                    => '&lt;br>%s',
+
+      '<div>%@</div>'             => '&lt;div>%s&lt;/div>',
+
+      '<span style="color:#45C1D0">%@</span>' => '&lt;span style="color:#45C1D0">%s&lt;/span>',
+
+      '<p>%@</p>'                 => '&lt;p>%s&lt;/p>',
+      '<p dir="ltr">%@</p>'       => '&lt;p dir="ltr">%s&lt;/p>',
 
       '<a href="target">link</a>'     => '<a href="target">link</a>',
       '<a href="target">"link"</a>'   => '<a href="target">\"link\"</a>',
       '<a href="target"></a>"out"'    => '<a href="target"></a>\"out\"',
       '<a href="http://url.com?param=1&param2=3&param3=%20">link</a>'   =>   '<a href="http://url.com?param=1&param2=3&param3=%20">link</a>',
 
-      '<p>escaped</p><![CDATA[]]>'                => '&lt;p>escaped&lt;/p><![CDATA[]]>',
-      '<![CDATA[]]><p>escaped</p>'                => '<![CDATA[]]>&lt;p>escaped&lt;/p>',
-      '<![CDATA[<p>unescaped</p>]]>'              => '<![CDATA[<p>unescaped</p>]]>',
-      '<![CDATA[<p>unescaped with %@</p>]]>'      => '<![CDATA[<p>unescaped with %s</p>]]>',
-      '<![CDATA[]]><![CDATA[<p>unescaped</p>]]>'  => '<![CDATA[]]><![CDATA[<p>unescaped</p>]]>',
+      '<q>escaped</q><![CDATA[]]>'                => '&lt;q>escaped&lt;/q><![CDATA[]]>',
+      '<![CDATA[]]><q>escaped</q>'                => '<![CDATA[]]>&lt;q>escaped&lt;/q>',
+      '<![CDATA[<q>unescaped</q>]]>'              => '<![CDATA[<q>unescaped</q>]]>',
+      '<![CDATA[<q>unescaped with %@</q>]]>'      => '<![CDATA[<q>unescaped with %s</q>]]>',
+      '<![CDATA[]]><![CDATA[<q>unescaped</q>]]>'  => '<![CDATA[]]><![CDATA[<q>unescaped</q>]]>',
 
       '<![CDATA[&]]>'  => '<![CDATA[&]]>',
       '<![CDATA[\']]>' => '<![CDATA[\']]>',
@@ -76,6 +145,11 @@ class TestAndroidFormatter < FormatterTest
       '<xliff:g>untouched</xliff:g>' => '<xliff:g>untouched</xliff:g>',
       '<xliff:g id="42">untouched</xliff:g>' => '<xliff:g id="42">untouched</xliff:g>',
       '<xliff:g id="1">first</xliff:g> inbetween <xliff:g id="2">second</xliff:g>' => '<xliff:g id="1">first</xliff:g> inbetween <xliff:g id="2">second</xliff:g>'
+    }
+    @escape_all_test_values = {
+      '<b>bold</b>'               => '&lt;b>bold&lt;/b>',
+      '<i>italic</i>'             => '&lt;i>italic&lt;/i>',
+      '<u>underline</u>'          => '&lt;u>underline&lt;/u>'      
     }
   end
 
@@ -101,6 +175,54 @@ class TestAndroidFormatter < FormatterTest
     assert_equal 'This is\n a string', @empty_twine_file.definitions_by_key["foo"].translations['en']
   end
 
+  def test_read_multiline_comment
+    content = <<-EOCONTENT
+      <?xml version="1.0" encoding="utf-8"?>
+      <resources>
+        <!-- multiline
+        comment -->
+        <string name="foo">This is
+     a string</string>
+      </resources>
+    EOCONTENT
+
+    io = StringIO.new(content)
+
+    @formatter.read io, 'en'
+
+    assert_equal 'multiline comment', @empty_twine_file.definitions_by_key["foo"].comment
+  end
+
+  def test_read_html_tags
+    content = <<-EOCONTENT
+      <?xml version="1.0" encoding="utf-8"?>
+      <resources>
+        <string name="foo">Hello, <b>BOLD</b></string>
+      </resources>
+    EOCONTENT
+
+    io = StringIO.new(content)
+
+    @formatter.read io, 'en'
+
+    assert_equal 'Hello, <b>BOLD</b>', @empty_twine_file.definitions_by_key["foo"].translations['en']
+  end
+
+  def test_double_quotes_are_not_modified
+    content = <<-EOCONTENT
+      <?xml version="1.0" encoding="utf-8"?>
+      <resources>
+        <string name="foo">Hello, <a href="http://www.foo.com">BOLD</a></string>
+      </resources>
+    EOCONTENT
+
+    io = StringIO.new(content)
+
+    @formatter.read io, 'en'
+
+    assert_equal 'Hello, <a href="http://www.foo.com">BOLD</a>', @empty_twine_file.definitions_by_key["foo"].translations['en']
+  end
+
   def test_set_translation_converts_leading_spaces
     @formatter.set_translation_for_key 'key1', 'en', "\u0020value"
     assert_equal ' value', @empty_twine_file.definitions_by_key['key1'].translations['en']
@@ -123,6 +245,11 @@ class TestAndroidFormatter < FormatterTest
 
   def test_set_translation_unescaping
     @escape_test_values.each do |expected, input|
+      @formatter.set_translation_for_key 'key1', 'en', input
+      assert_equal expected, @empty_twine_file.definitions_by_key['key1'].translations['en']
+    end
+
+    @escape_all_test_values.each do |expected, input|
       @formatter.set_translation_for_key 'key1', 'en', input
       assert_equal expected, @empty_twine_file.definitions_by_key['key1'].translations['en']
     end
@@ -154,6 +281,11 @@ class TestAndroidFormatter < FormatterTest
     @escape_test_values.each do |input, expected|
       assert_equal expected, @formatter.format_value(input)
     end
+
+    @formatter.options.merge!({ escape_all_tags: true })
+    @escape_all_test_values.each do |input, expected|
+      assert_equal expected, @formatter.format_value(input)
+    end
   end
 
   def test_format_value_escapes_non_resource_identifier_at_signs
@@ -165,8 +297,24 @@ class TestAndroidFormatter < FormatterTest
     assert_equal identifier, @formatter.format_value(identifier)
   end
 
+  def test_deducts_language_from_filename
+    language = KNOWN_LANGUAGES.sample
+    assert_equal language, @formatter.determine_language_given_path("#{language}.xml")
+  end
+
+  def test_recognize_every_twine_language_from_filename
+    twine_file = build_twine_file "not-a-lang-code" do
+      add_section "Section" do
+        add_definition key: "value"
+      end
+    end
+
+    @formatter.twine_file = twine_file
+    assert_equal "not-a-lang-code", @formatter.determine_language_given_path("not-a-lang-code.xml")
+  end
+
   def test_deducts_language_from_resource_folder
-    language = %w(en de fr).sample
+    language = KNOWN_LANGUAGES.sample
     assert_equal language, @formatter.determine_language_given_path("res/values-#{language}")
   end
 
@@ -185,6 +333,13 @@ class TestAndroidFormatter < FormatterTest
   def test_output_path_with_region
     assert_equal 'values-en-rGB', @formatter.output_path_for_language('en-GB')
   end
+
+  def test_output_path_respects_default_lang
+    @formatter.twine_file.language_codes.concat KNOWN_LANGUAGES
+    non_default_language = KNOWN_LANGUAGES[1..-1].sample
+    assert_equal 'values', @formatter.output_path_for_language(KNOWN_LANGUAGES[0])
+    assert_equal "values-#{non_default_language}", @formatter.output_path_for_language(non_default_language)
+  end
 end
 
 class TestAppleFormatter < FormatterTest
@@ -196,6 +351,22 @@ class TestAppleFormatter < FormatterTest
     @formatter.read content_io('formatter_apple.strings'), 'en'
 
     assert_file_contents_read_correctly
+  end
+
+  def test_deducts_language_from_filename
+    language = KNOWN_LANGUAGES.sample
+    assert_equal language, @formatter.determine_language_given_path("#{language}.strings")
+  end
+
+  def test_recognize_every_twine_language_from_filename
+    twine_file = build_twine_file "not-a-lang-code" do
+      add_section "Section" do
+        add_definition key: "value"
+      end
+    end
+
+    @formatter.twine_file = twine_file
+    assert_equal "not-a-lang-code", @formatter.determine_language_given_path("not-a-lang-code.strings")
   end
 
   def test_deducts_language_from_resource_folder
@@ -279,6 +450,12 @@ class TestJQueryFormatter < FormatterTest
     assert_translations_read_correctly
   end
 
+  def test_read_format_nested
+    @formatter.read content_io('formatter_jquery_nested.json'), 'en'
+
+    assert_nested_translations_read_correctly
+  end
+
   def test_format_file
     formatter = Twine::Formatters::JQuery.new
     formatter.twine_file = @twine_file
@@ -301,6 +478,21 @@ class TestJQueryFormatter < FormatterTest
 
   def test_format_value_with_newline
     assert_equal "value\nwith\nline\nbreaks", @formatter.format_value("value\nwith\nline\nbreaks")
+  end
+
+  def test_deducts_language_from_filename
+    language = KNOWN_LANGUAGES.sample
+    assert_equal language, @formatter.determine_language_given_path("#{language}.json")
+  end
+
+  def test_deducts_language_from_extended_filename
+    language = KNOWN_LANGUAGES.sample
+    assert_equal language, @formatter.determine_language_given_path("something-#{language}.json")
+  end
+
+  def test_deducts_language_from_path
+    language = %w(en-GB de fr).sample
+    assert_equal language, @formatter.determine_language_given_path("/output/#{language}/#{@formatter.default_file_name}")
   end
 end
 
@@ -331,6 +523,21 @@ class TestGettextFormatter < FormatterTest
     language = "en-GB"
     assert_equal language, @formatter.determine_language_given_path("#{language}.po")
   end
+
+  def test_deducts_language_from_path
+    language = %w(en-GB de fr).sample
+    assert_equal language, @formatter.determine_language_given_path("/output/#{language}/#{@formatter.default_file_name}")
+  end
+
+  def test_quoted_strings
+    formatter = Twine::Formatters::Gettext.new
+    formatter.twine_file = build_twine_file "not-a-lang-code" do
+      add_section "Section" do
+        add_definition key: "foo \"bar\" baz"
+      end
+    end
+    assert_equal content('formatter_gettext_quotes.po'), formatter.format_file('en')
+  end
 end
 
 class TestTizenFormatter < FormatterTest
@@ -351,7 +558,6 @@ class TestTizenFormatter < FormatterTest
     formatter.twine_file = @twine_file
     assert_equal content('formatter_tizen.xml'), formatter.format_file('en')
   end
-
 end
 
 class TestDjangoFormatter < FormatterTest
@@ -374,6 +580,24 @@ class TestDjangoFormatter < FormatterTest
   def test_deducts_language_and_region
     language = "en-GB"
     assert_equal language, @formatter.determine_language_given_path("#{language}.po")
+  end
+
+  def test_deducts_language_from_path
+    language = %w(en-GB de fr).sample
+    assert_equal language, @formatter.determine_language_given_path("/output/#{language}/#{@formatter.default_file_name}")
+  end
+
+  def test_ignores_commented_out_strings
+    content = <<-EOCONTENT
+      #~ msgid "foo"
+      #~ msgstr "This should be ignored"
+    EOCONTENT
+
+    io = StringIO.new(content)
+
+    @formatter.read io, 'en'
+
+    assert_nil @empty_twine_file.definitions_by_key["foo"]
   end
 end
 
@@ -405,10 +629,10 @@ class TestFlashFormatter < FormatterTest
 
   def test_deducts_language_from_resource_folder
     language = %w(en de fr).sample
-    assert_equal language, @formatter.determine_language_given_path("locale/#{language}")
+    assert_equal language, @formatter.determine_language_given_path("locale/#{language}/#{@formatter.default_file_name}")
   end
 
   def test_deducts_language_and_region_from_resource_folder
-    assert_equal 'de-AT', @formatter.determine_language_given_path("locale/de-AT")
+    assert_equal 'de-AT', @formatter.determine_language_given_path("locale/de-AT/#{@formatter.default_file_name}")
   end
 end