+++ test/test_generate_localization_archive.rb
@@ -1,8 +1,7 @@
 require 'command_test'
 
 class TestGenerateLocalizationArchive < CommandTest
-  def new_runner(twine_file = nil)
-    options = {}
+  def new_runner(twine_file = nil, options = {})
     options[:output_path] = @output_path
     options[:format] = 'apple'
 
@@ -20,7 +19,7 @@ class TestGenerateLocalizationArchive < CommandTest
   def test_generates_zip_file
     new_runner.generate_localization_archive
 
-    assert File.exists?(@output_path), "zip file should exist"
+    assert File.exist?(@output_path), "zip file should exist"
   end
 
   def test_zip_file_structure
@@ -45,7 +44,13 @@ class TestGenerateLocalizationArchive < CommandTest
   def test_prints_empty_file_warnings
     empty_twine_file = build_twine_file('en') {}
     new_runner(empty_twine_file).generate_localization_archive
-    assert_match "Skipping file", Twine::stderr.string
+    assert_match "Skipping file", Twine::stdout.string
+  end
+
+  def test_does_not_print_empty_file_warnings_if_quite
+    empty_twine_file = build_twine_file('en') {}
+    new_runner(empty_twine_file, quite: true).generate_localization_archive
+    refute_match "Skipping file", Twine::stdout.string
   end
 
   class TestValidate < CommandTest